custom-elements/instantiating-custom-elements/extensions-to-document-interface/create-element-test-003.html

<!DOCTYPE html>
<html>
<head>
<title>If typeExtension is unresolved when createElement called then local name is a type</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="If an element definition with matching localName, namespace, and TYPE is not registered with token's document, set TYPE to localName">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#extensions-to-document-interface-to-instantiate">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test (function () {
    for (var i = 0; i < HTML5_ELEMENTS.length; i++) {
        var element = HTML5_ELEMENTS[i];
        
        var doc = newHTMLDocument();
        
        var obj = doc.createElement(element);
        
        var name = 'x-1-' + element;
        
        var proto = Object.create(obj.constructor.prototype);
        
        var customElement = doc.createElement(element, name);
            
        assert_equals(customElement.constructor.prototype.constructor.name, obj.constructor.name, 
                'Wrong custom element type before registration of the custom element type');
        
        doc.register(name, {prototype: proto, extends: element});
        
        assert_equals(customElement.constructor.prototype.constructor.name, name, 
                'Wrong custom element type after registration of the custom element type');
    }    
}, 'If typeExtension presents in createElement call then type is a type extension');


test (function () {
    for (var i = 0; i < HTML5_ELEMENTS.length; i++) {
        var element = HTML5_ELEMENTS[i];
        
        var doc = newHTMLDocument();
        
        var obj = doc.createElement(element);
        
        var name = 'x-2-' + element;
        
        var proto = Object.create(obj.constructor.prototype);
        
        var customElement = doc.createElementNS(HTML_NAMESPACE, element, name);
            
        assert_equals(customElement.constructor.prototype.constructor.name, obj.constructor.name, 
                'Wrong custom element type before registration of the custom element type');
        
        doc.register(name, {prototype: proto, extends: element});
        
        assert_equals(customElement.constructor.prototype.constructor.name, name, 
                'Wrong custom element type after registration of the custom element type');
    }    
}, 'If typeExtension presents in createElementNS call then type is a type extension');

</script>
</body>
</html>
